remove eventmanager change
