+++ src/main/java/Bots/commands/general/CommandHelp.java
@@ -14,7 +14,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 
-import static Bots.managers.EventManager.registerButtonInteraction;
 import static Bots.managers.LocaleManager.managerLocalise;
 import static Bots.Main.*;
