+++ src/main/java/Bots/commands/music/CommandQueue.java
@@ -18,7 +18,6 @@ import java.util.*;
 import java.util.concurrent.BlockingQueue;
 
 import static Bots.EmbedHelper.toTimestamp;
-import static Bots.managers.EventManager.registerButtonInteraction;
 import static Bots.managers.LocaleManager.managerLocalise;
 import static Bots.Main.*;
